refactor: separate tests into smaller units + files
